throw blog header and search together
